requirements/requirements_for_dev.txt
flake8==6.0.0
black==23.1.0
pyflakes==3.0.1
isort==5.12.0
flake8-quotes==3.3.2
flake8-import-order==0.18.2
django_debug_toolbar==3.8.1